Move ResourceLimiter to separate file
